==> notebooks/synthetic_data/Step1_generate_data.ipynb <==
As the first step of this tutorial, we will generate some synthetic data to train a model.

This code generates solid color swatches and striped color swatches.

You can see the code in `generate_swatches.py`.

Note: this code will create a directory called `data` filled with png files.

%load_ext autoreload

%autoreload 2

import pandas as pd
from sklearn.model_selection import train_test_split

from generate_swatches import generate_color_swatches, generate_pattern_swatches

Generate Synthetic Data
===

generate_color_swatches()

generate_pattern_swatches()

Note: if you changed the default storage location above, you will need to alter the next two lines.

COLOR_DATASET_PATH = 'data/color_swatches'

PATTERN_DATASET_PATH = 'data/pattern_swatches'

Format color data
===

Now we will put the data in the format needed for Octopod.

color_df = pd.read_csv(f'{COLOR_DATASET_PATH}/color_dataset.csv')

color_df.head()

You can see the generated csv `color_dataset.csv` contains four columns:
- complex-color: the value we will use for the text
- simple_color: the value we will try to predict
- rgb: the raw RGB value used to generate the swatch
- image_locs: the location of where each image is stored

We set `simple_color` to be a pandas category type

color_df['simple_color'] = color_df['simple_color'].astype('category')

We create a new column called `simple_color_cat` that contains the category to predict as an integer.
This is an important step because Octopod does not natively handle categories as text

color_df['simple_color_cat'] = color_df['simple_color'].cat.codes

# Category mapping
{label: i for i, label in enumerate(color_df['simple_color'].cat.categories)}

Note: you will need to save this mapping for later so that you can use your model to make predictions.
Since this is a simple model, we won't save it, but for a real project, we would store it somewhere.

color_df.head()

Now our color data is in the right format for Octopod so we'll split it and save it again.

color_train_df, color_valid_df = train_test_split(color_df, train_size=0.75, random_state=17)

color_train_df.to_csv(f'{COLOR_DATASET_PATH}/color_train.csv', index=False)

color_valid_df.to_csv(f'{COLOR_DATASET_PATH}/color_valid.csv', index=False)

Format pattern data
===

pattern_df = pd.read_csv(f'{PATTERN_DATASET_PATH}/pattern_dataset.csv')

pattern_df.head()

You can see the generated csv `pattern_dataset.csv` contains two columns:
- pattern_type: the value we will try to predict
- image_locs: the location of where each image is stored

We set `pattern_type` to be a pandas category type

pattern_df['pattern_type'] = pattern_df['pattern_type'].astype('category')

We create a new column called `pattern_type_cat` that contains the category to predict as an integer.
This is an important step because Octopod does not natively handle categories as text

pattern_df['pattern_type_cat'] = pattern_df['pattern_type'].cat.codes

# Category mapping
{label: i for i, label in enumerate(pattern_df['pattern_type'].cat.categories)}

Note: you will need to save this mapping for later so that you can use your model to make predictions.
Since this is a simple model, we won't save it, but for a real project, we would store it somewhere.

Since we don't have any text for our pattern data, we will fake it by just using text from the filenames.
Note: this would be a terrible idea for a real model since the target variable is contained in the file name.

pattern_df['fake_text'] = [i.replace('/', ' ') for i in pattern_df['image_locs']]

pattern_df.loc[0]['image_locs']

pattern_df.loc[0]['fake_text']

pattern_df.head()

Now our pattern data is in the right format for Octopod so we'll split it and save it again.

pattern_train_df, pattern_valid_df = train_test_split(pattern_df, train_size=0.75, random_state=17)

pattern_train_df.to_csv(f'{PATTERN_DATASET_PATH}/pattern_train.csv', index=False)

pattern_valid_df.to_csv(f'{PATTERN_DATASET_PATH}/pattern_valid.csv', index=False)

Now that we have some data, move onto the notebook labeled `Step2_train_image_model.ipynb`.
